change ops patch to patch_common dir
