+++ vllm_ascend/patch/worker/patch_common/__init__.py
@@ -18,8 +18,7 @@
 from vllm.triton_utils import HAS_TRITON
 
 if HAS_TRITON:
-    import vllm_ascend.ops.casual_conv1d
-    import vllm_ascend.ops.sigmoid_gating
+    import vllm_ascend.patch.worker.patch_common.patch_triton
 
 import vllm_ascend.patch.worker.patch_common.patch_distributed  # noqa
 import vllm_ascend.patch.worker.patch_common.patch_logits  # noqa
